Updated to dmd2.063, switched to std.process and abandoned std.concurrency.
dmd2.063 finally brought the new std.process into phobos, so it
was imperative to use it ASAP. It works very nicely.

There was hope for a while that the problems with Variant had also
been addressed, but alas it was not to be. This precipitated replacing
std.concurrency with a VERY simple home-grown version that turns out
to be a whole lot more satisfying. It is slightly more cumbersome
to use, but the compiler can now provide heaps of compile-time checking,
and it is very easy to understand what is going on - quite the opposite
of std.concurrency and std.variant.