Fixed a bug re circular includes.
This change set adds an explicit check for circular includes
as includes are scanned for. This fixes a problem introduced recently
when processing of the consequences of includes was delayed until
all the downstream includes are known.

Also fixed/simplified a few things with inter-thread communication
and error handling that became obvious during testing of the above.

Things are starting to look much more solid now.