Split bub.d up into several files and added Makefile plus bub.cfg.
It was time to split bub.d up into smaller files, and this was a
good opportunity to show how much easier bub is to use than make.
Note that I am no make expert, and there are ways to write a much
cleaner Makefile, but my experience is that it is always far too hard
to do simple things like: put object files here, binaries there, etc.

The Makefile allows bub to be built sensibly, speeding up
compiles of bub during development.

The bub.cfg file allows bub itself to serve as a simple-as-possible
getting-started example. It doesn't show off any of bub's cool
features, but it does show that bub is easy to use.