Add ARCHITECTURE-dependent blocks in Bubfiles.
If an ARCHITECTURE variable is defined, then a block of code like this
in a Bubfile:

  [value] { some-text }

resolves to some-text if value is one of ARCHITECTURE's values, and
empty-string if not.

This allows architecture-dependent features/implementations to be supported
in a single source tree, enabled via different configuration files or
different modes.

Example use cases:

* Contain different packages.

* Incorporate different private files in a library.

* Omit some tests.

Note that architecture-dependent statements should be used very sparingly,
and their use should be concentrated in a very small number of low-level
packages. Use conditional compilation instead where feasible.