Separate content-specific config information into bundle files.
In large projects that reuse code, it is better to reuse information in
bundle files in the reuse repos rather than copy into into each project's
bub.cfg file.

A specific use case this is intended to support is where you want to
build 3rd-party libraries with bub as part of your project rather than
pulling them in as system libraries.