Rework calculation of what libraries are depended on to be more efficient.
Previously, the implementation was correct, but very computationally
expensive. This commit simplifies the implementation and improves
its performance considerably, at the cost of adding extra
dependencies from dynamic libraries to static libraries that
aren't strictly necessary.

The main result is that the wall-clock time taken for bub to work
through a project of several thousand files (source and built) that are
already up to date has dropped from about 0.7 seconds to about 0.3
seconds. If you are changing files at the top of the dependency pile,
this reduction in the delay before things start to happen is very noticeable.