Change processing of already-clean files to limit stack depth.
Previously, already-clean files were immediately marked as up-to-date,
which had knock-on consequences that resulted in other clean files
also being marked as up-to-date, which drove an unbounded recursion.

This commit removes the recursion by putting all ready actions
on the work queue even if they don't need to be given to a worker,
and discarding them as appropriate as they come off the queue.

While here, the implementation of blocking actions that need to
wait for generated files was reworked to make it simpler and faster.