Add an optional second layer to modes to facilitate complex mode composition.
When you have multiple dimensions of interest, defining each combination
of those concerns as a mode leads to massive duplication.

Specifying multiple modes on the command-line would be problematic
because incompatible combinations would be difficult to prevent.

The approach taken here is to define several mode fragments, and combine
appropriate sets of them into modes. If no modes are defined, the
fragments are promoted to modes.