Change test failure output to prompt an editor to open the failure results file.
The failure output is now written to a file in priv, and its content
plus a magic line are written to stderr. The magic line is intended to
prompt an editor to open the failure file, similarly to how such an
editor would open the soure file containing the first error from a
build run.

Tested with vim.