Create/update files-of-interest and include-paths files at the start of a build.
These files are intended to be of use for developers who use an IDE that
doesn't use compile_commands.json. The idea is that such users will watch for
updates to these files, and use their content as an input for refreshing their
IDE's project file or similar.

The files of interest are deemed to be any file that is part of the build
that isn't an object file, static library, dynamic library, executable or
test result file. Quite a few of these may still be uninteresting to the IDE, so
further project/IDE/user specific pruning will probably be useful.

The include paths are the union of all the -iquote, -isystem or -I paths used
in the whole project. While this will enable all header files to be found, there
is always the danger that name collisions may occur in system header files.
The only way to address that is to use source-file-specific include paths,
which brings us back to compile_commands.json as the preferred approach if
that is an option.

http://reviewboard.acres.com.au/r/2303/