Execute commands via shell.
Previously bub was treating commands derived from bub.cfg as literal
commands and splitting up the arguments based on whitespace. This was
unconventional.

Now it passes the commands to the underlying shell to interpret.
Consequently, special characters (such as quotes and backslashes) must
be escaped in bub.cfg (in the usual shell way).

Also, fix the escaping of commands in compile_commands.json.

http://reviewboard.acres.com.au/r/2304/