Add support for companion files as a side effect of compilation.
This change is intended to support things like code coverage analysis, by
declaring that some rules can produce "companion" files that are
not part of the build as such, but will not be deleted by bub.

The example of most interest is:

.c .obj .gcno = <the-command>

which says: this is how to produce an object file from a .c file, and
it *may* produce a .gcno file that should not be deleted.

While here, removed the suppression of duplicate values in variable declarations
because some compilers use command-line arguments that contain embedded spaces.
The examples that triggered this change are clang's:

    -Xclang -coverage-cfg-checksum
    -Xclang -coverage-no-function-names-in-data

http://reviewboard.acres.com.au/r/2319/