Fix a possible infinite loop in flushPkgDepends.
An infinie loop in a foreach body in planner/flushPkgDepends has been
observed.

Studying the code, it appears that the problem was likely to be caused by
mutating a copy of an associative array value rather than a reference to it.

This change makes the mutated foreach values "ref", and adds a fatal() if
a pass through the loop unexpectedly fails to find any package with no
remaining dependencies.

Over the shoulder review by Clive.

noreview