Fix segfault in case of depending on header files in executables.
Fixed a bug in handling the case where an executable depends on a header
file that is contained by another executable. Also improved the error
message.

Note: Executables should not usually contain header files, but it is not
prevented because in cases where the header is only used by one
executable it is more convenient than putting the header in a separate
library.

noreview