Check whether the build directory needs re-configuring.
The primary purpose of this change is to verify that
src/Bubfile is consistent with the project's bub.cfg and
the bundle files it references. Such an inconsistency can
otherwise lead to difficulty in identifying build problems.

If the check fails, so does the build - with a message
recommending that bub-config be re-run.

OTS review by Dave.

noreview