Discard dependencies outside the project.
Raw D dependency files don't distinguish between system and
project dependencies, and the existing test for a file being within
the build directory doesn't discard system files present in the build
directory (say in a 3rd-party subdirectory).

This change narrows the definition of in-project dependencies to just
files in src, obs or directly in the build directory.

OTS review by grahams.

noreview