Remove TMP_PATH considerations from the run script, and set TMPDIR when running tests.
The run script should never have been creating and destroying directories.
It is up to the process being run to decide on a location for its temporary
files and directories, which will usually be /tmp.

When running tests, bub sets TMPDIR (which is the usual convention
rather than TMP_PATH), creates that directory before running the test,
and removes it on test success.

REUSE-159

http://reviewboard.acres.com.au/r/2917/