Rename ARCHITECTURE to CONDITIONALS and add late resolution of its value.
ARCHITECTURE was always a bad name.

CONDITIONALS implies a connection to conditional Bubfile content.

Late resolution of the value of CONDITIONALS allows its value to
be mode-dependent, removing the need for a different configuration
file for each set of CONDITIONALS values that are needed in a project.

OTS review by Dave.

noreview