Remove an action's cached dependences before issuing the action to a worker
If bub is killed with SIGKILL, it and its children do not get an
opportunity to clean up properly. This can result in partially
created build artifacts and stale dependencies.

This commit removes cached dependencies for the built files
an action will produce *before* issuing the action to a worker thread.
The cached dependencies are updated on successful completion of
the action as before. Should a build be terminated with a SIGKILL,
the in-progress actions will have no cached dependencies, which causes
the action to be re-issued regardless of the possible presence of
incomplete built files.

BUB-16

noreview