bub: Add a command-line flag to cause built files to be touched after being updated
The encfs filesystem corrupts file modification times when files
or directories are moved. Specifically, when a build action writes to
a temporary file and renames the file on completion, the file's mtime's
fractional seconds are set to zero. This causes files to be rebuilt
unnecessarily on subsequent builds.

This commit adds the --touch flag, which causes any updated built file's
modification and access times to be updated, working around this
problem.

BUB-20

noreview