Update all associated files on Bubfile changes
Previously, a change to a Bubfile would force the rebuild of all
built files defined in the Bubfile. However, this did not cause
all the relevant higher-level files to be rebuilt, so some changes
(such as making a header file private) did not result in the
appropriate build error.

A change in a Bubfile now effectively touches all the source
files mentioned in the Bubfile, resolving the problem.

BUB-22

http://reviewboard.acres.com.au/r/3563/